llsf_sps: rename files to match style of the rest of the project
We do not use CamelCase but names_with_underlines.
