+++ src/libs/mps_comm/ring_station.cpp
@@ -1,5 +1,5 @@
-#include "RingStation.h"
-#include "MPSIoMapping.h"
+#include "ring_station.h"
+#include "mps_io_mapping.h"
 #include <iostream>
 
 namespace llsfrb {
